Tests/LibWeb/Text/expected/navigable-element-mouseleave.txt

Red Background in iframe: true
Red Background in iframe: false
